--- a/bench/.gitignore
+++ b/bench/.gitignore
@@ -1,0 +1,3 @@
+/corpus/
+/results/
+/worktree/