## Complete traversal

### linux v6.5 (86380 files)

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main bench/corpus/linux -false` | 24.7 ± 5.1 | 18.4 | 35.4 | 1.00 |
| `find bench/corpus/linux -false` | 86.0 ± 0.3 | 85.5 | 86.5 | 3.48 ± 0.71 |
| `fd -u '^$' bench/corpus/linux` | 171.3 ± 42.6 | 123.6 | 220.0 | 6.93 ± 2.23 |

### rust 1.72.1 (192714 files)

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main bench/corpus/rust -false` | 57.2 ± 5.8 | 49.3 | 69.8 | 1.00 |
| `find bench/corpus/rust -false` | 276.1 ± 0.5 | 275.4 | 277.0 | 4.83 ± 0.49 |
| `fd -u '^$' bench/corpus/rust` | 292.0 ± 44.1 | 240.0 | 343.6 | 5.10 ± 0.93 |

### chromium 119.0.6036.2 (2119292 files)

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main bench/corpus/chromium -false` | 531.8 ± 13.1 | 511.3 | 552.1 | 1.00 |
| `find bench/corpus/chromium -false` | 2925.0 ± 27.2 | 2865.7 | 2963.6 | 5.50 ± 0.15 |
| `fd -u '^$' bench/corpus/chromium` | 2427.3 ± 28.1 | 2348.3 | 2443.3 | 4.56 ± 0.12 |

## Early termination

### chromium 119.0.6036.2 (depth 24)

#### Depth 2

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main bench/corpus/chromium -name $(shuf -n1 $FILES) -print -quit` | 8.5 ± 2.1 | 4.2 | 12.6 | 1.00 |
| `find bench/corpus/chromium -name $(shuf -n1 $FILES) -print -quit` | 3293.6 ± 86.1 | 3185.5 | 3450.9 | 385.40 ± 96.23 |
| `fd -usg1 $(shuf -n1 $FILES) bench/corpus/chromium` | 235.2 ± 236.3 | 85.1 | 1063.0 | 27.52 ± 28.48 |

#### Depth 4

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main bench/corpus/chromium -name $(shuf -n1 $FILES) -print -quit` | 32.0 ± 11.7 | 14.3 | 48.1 | 1.00 |
| `find bench/corpus/chromium -name $(shuf -n1 $FILES) -print -quit` | 2051.9 ± 1526.3 | 70.7 | 3488.7 | 64.22 ± 53.27 |
| `fd -usg1 $(shuf -n1 $FILES) bench/corpus/chromium` | 1136.7 ± 971.5 | 63.7 | 2431.9 | 35.57 ± 33.09 |

#### Depth 8

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main bench/corpus/chromium -name $(shuf -n1 $FILES) -print -quit` | 255.4 ± 16.8 | 233.1 | 281.7 | 1.00 |
| `find bench/corpus/chromium -name $(shuf -n1 $FILES) -print -quit` | 2724.7 ± 689.5 | 1528.7 | 3244.7 | 10.67 ± 2.79 |
| `fd -usg1 $(shuf -n1 $FILES) bench/corpus/chromium` | 582.3 ± 478.3 | 134.4 | 1792.1 | 2.28 ± 1.88 |

#### Depth 16

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main bench/corpus/chromium -name $(shuf -n1 $FILES) -print -quit` | 540.3 ± 18.5 | 519.3 | 564.0 | 1.00 |
| `find bench/corpus/chromium -name $(shuf -n1 $FILES) -print -quit` | 1143.1 ± 1099.2 | 91.7 | 2806.3 | 2.12 ± 2.04 |
| `fd -usg1 $(shuf -n1 $FILES) bench/corpus/chromium` | 1131.6 ± 681.1 | 176.2 | 2405.3 | 2.09 ± 1.26 |

## Printing paths

### Without colors

#### linux v6.5

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main bench/corpus/linux` | 35.2 ± 9.5 | 26.1 | 50.8 | 1.00 |
| `find bench/corpus/linux` | 100.7 ± 11.8 | 93.1 | 132.2 | 2.86 ± 0.84 |
| `fd -u --search-path bench/corpus/linux` | 195.1 ± 28.4 | 130.8 | 215.1 | 5.55 ± 1.70 |

### With colors

#### linux v6.5

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main bench/corpus/linux -color` | 199.4 ± 12.4 | 187.4 | 218.0 | 1.00 |
| `fd -u --search-path bench/corpus/linux --color=always` | 207.9 ± 21.8 | 138.4 | 221.4 | 1.04 ± 0.13 |

## Search strategies

### linux v6.5

#### bfs-main

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `bfs-main -S bfs bench/corpus/linux` | 34.2 ± 11.7 | 21.0 | 50.7 | 1.00 |
| `bfs-main -S dfs bench/corpus/linux` | 35.5 ± 9.4 | 28.8 | 53.3 | 1.04 ± 0.45 |
| `bfs-main -S ids bench/corpus/linux` | 133.6 ± 14.4 | 110.3 | 152.8 | 3.90 ± 1.40 |
| `bfs-main -S eds bench/corpus/linux` | 58.4 ± 9.5 | 43.6 | 74.1 | 1.71 ± 0.65 |

## Details

### Versions

```console
$ bfs-main --version | head -n1
bfs 3.0.2-14-g52de184b
$ find --version | head -n1
find (GNU findutils) 4.9.0
$ fd --version
fd 8.7.0
```